Use new packages for layout and module models
