+++ src/shogun2-core/shogun2-dao/src/main/java/de/terrestris/shogun2/dao/ModuleDao.java
@@ -2,7 +2,7 @@ package de.terrestris.shogun2.dao;
 
 import org.springframework.stereotype.Repository;
 
-import de.terrestris.shogun2.model.Module;
+import de.terrestris.shogun2.model.module.Module;
 
 @Repository
 public class ModuleDao extends GenericHibernateDao<Module, Integer> {
