+++ src/shogun2-core/shogun2-model/src/main/java/de/terrestris/shogun2/model/layout/AbsoluteLayout.java
@@ -1,8 +1,9 @@
 /**
  * 
  */
-package de.terrestris.shogun2.model;
+package de.terrestris.shogun2.model.layout;
 
+import java.awt.Point;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,12 +21,14 @@ import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
+import de.terrestris.shogun2.model.module.CompositeModule;
+
 /**
- * This class is the representation of an border layout, where components are
- * anchored in (predefined) regions, which are stored in the {@link #regions}
+ * This class is the representation of an absolute layout, where components are
+ * anchored in absolute positions, which are stored in the {@link #coords}
  * property.
  * 
- * The order of the {@link #regions} should match the order of the corresponding
+ * The order of the {@link #coords} should match the order of the corresponding
  * {@link CompositeModule#getSubModules()}.
  * 
  * @author Nils Bühner
@@ -33,7 +36,7 @@ import org.apache.commons.lang3.builder.ToStringStyle;
  */
 @Table
 @Entity
-public class BorderLayout extends Layout {
+public class AbsoluteLayout extends Layout {
 
 	/**
 	 * 
@@ -44,32 +47,32 @@ public class BorderLayout extends Layout {
 	 * Explicitly adding the default constructor as this is important, e.g. for
 	 * Hibernate: http://goo.gl/3Cr1pw
 	 */
-	public BorderLayout() {
-		this.setType("border");
+	public AbsoluteLayout() {
+		this.setType("absolute");
 	}
 
 	/**
 	 * 
 	 */
 	@ElementCollection(fetch = FetchType.EAGER)
-	@CollectionTable(name = "BORDERLAYOUT_REGIONS", joinColumns = @JoinColumn(name = "LAYOUT_ID") )
-	@Column(name = "REGION")
+	@CollectionTable(name = "ABSOLUTELAYOUT_COORDS", joinColumns = @JoinColumn(name = "LAYOUT_ID") )
+	@Column(name = "COORD")
 	@OrderColumn(name = "INDEX")
-	private List<String> regions = new ArrayList<String>();
+	private List<Point> coords = new ArrayList<Point>();
 
 	/**
-	 * @return the regions
+	 * @return the coords
 	 */
-	public List<String> getRegions() {
-		return regions;
+	public List<Point> getCoords() {
+		return coords;
 	}
 
 	/**
-	 * @param regions
-	 *            the regions to set
+	 * @param coords
+	 *            the coords to set
 	 */
-	public void setRegions(List<String> regions) {
-		this.regions = regions;
+	public void setCoords(List<Point> coords) {
+		this.coords = coords;
 	}
 
 	/**
@@ -82,7 +85,7 @@ public class BorderLayout extends Layout {
 	 */
 	public int hashCode() {
 		// two randomly chosen prime numbers
-		return new HashCodeBuilder(53, 23).appendSuper(super.hashCode()).append(getRegions()).toHashCode();
+		return new HashCodeBuilder(11, 13).appendSuper(super.hashCode()).append(getCoords()).toHashCode();
 	}
 
 	/**
@@ -94,11 +97,11 @@ public class BorderLayout extends Layout {
 	 *      when using ORM like Hibernate
 	 */
 	public boolean equals(Object obj) {
-		if (!(obj instanceof BorderLayout))
+		if (!(obj instanceof AbsoluteLayout))
 			return false;
-		BorderLayout other = (BorderLayout) obj;
+		AbsoluteLayout other = (AbsoluteLayout) obj;
 
-		return new EqualsBuilder().appendSuper(super.equals(other)).append(getRegions(), other.getRegions()).isEquals();
+		return new EqualsBuilder().appendSuper(super.equals(other)).append(getCoords(), other.getCoords()).isEquals();
 	}
 
 	/**
@@ -106,7 +109,7 @@ public class BorderLayout extends Layout {
 	 */
 	public String toString() {
 		return new ToStringBuilder(this, ToStringStyle.DEFAULT_STYLE).appendSuper(super.toString())
-				.append("regions", getRegions()).toString();
+				.append("coords", getCoords()).toString();
 	}
 
 }
