+++ src/shogun2-core/shogun2-model/src/main/java/de/terrestris/shogun2/model/module/LayerTree.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package de.terrestris.shogun2.model;
+package de.terrestris.shogun2.model.module;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
