+++ src/shogun2-core/shogun2-service/src/main/java/de/terrestris/shogun2/service/ModuleService.java
@@ -2,7 +2,7 @@ package de.terrestris.shogun2.service;
 
 import org.springframework.stereotype.Service;
 
-import de.terrestris.shogun2.model.Module;
+import de.terrestris.shogun2.model.module.Module;
 
 /**
  * Service class for the {@link Module} model.
