+++ src/shogun2-web/src/main/java/de/terrestris/shogun2/web/LayoutController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import de.terrestris.shogun2.model.Layout;
+import de.terrestris.shogun2.model.layout.Layout;
 import de.terrestris.shogun2.service.LayoutService;
 
 /**
